Query Single-Cell RNA-Seq Dataset

This dataset contains the processed query dataset from the HeOrganAtlas dataset for Marrow tissue. It has been preprocessed to include log-normalized counts, specific metadata columns, annotations based on SingleR cell type scoring, and PCA, t-SNE, and UMAP results.

query_data

Format

An object of class SingleCellExperiment with 392 rows and 503 columns.

Source

The HeOrganAtlas dataset, available through the scRNAseq package.

Details

This dataset underwent the following steps:

Loads the HeOrganAtlas dataset specifically for Marrow tissue from the scRNAseq package.

Divides the loaded dataset into a query dataset used for downstream analysis.

Performs log normalization on the query dataset using the function logNormCounts from the scuttle package.

Selects specific columns (percent_mito, expert_annotation) from the cell metadata for downstream analysis.

Adds SingleR annotations (SingleR_annotation) and annotation scores (annotation_scores) to the query dataset using the function SingleR from the SingleR package.

Computes AUC gene set scores using the function AUCell_calcAUC from the AUCell package and adds these scores to the query dataset.

Selects highly variable genes (HVGs) using the function getTopHVGs from the scran package on the query dataset.

Intersects the highly variable genes between the query and reference datasets to obtain common genes for analysis.

Performs Principal Component Analysis (PCA) on the query dataset using the function runPCA from the scater package.

Performs t-Distributed Stochastic Neighbor Embedding (t-SNE) on the query dataset using the function runTSNE from the scater package.

Performs Uniform Manifold Approximation and Projection (UMAP) on the query dataset using the function runUMAP from the scater package.

References

He, et al. (2020). HeOrganAtlas: a comprehensive human organ atlas based on single-cell RNA sequencing.

See also

Examples

# Load and explore the query dataset
data("query_data")