Convert Specified Columns to Character in SingleCellExperiment Objects

This function converts specified columns in the colData of a SingleCellExperiment object to character type. It checks that the specified columns exist and only performs conversion when necessary (i.e., when columns are not already character type).

convertColumnsToCharacter(sce_object, convert_cols)

Arguments

sce_object

A SingleCellExperiment object containing single-cell data.

convert_cols

A character vector specifying the column names in colData to convert to character type. All specified columns must exist in the colData.

Value

A SingleCellExperiment object with the specified columns converted to character type in the colData.

Details

The function performs the following operations:

Validates that the input is a SingleCellExperiment object.

Checks that all specified columns exist in the colData of the object.

Converts each specified column to character type if it is not already character.

Returns the modified SingleCellExperiment object.

If all specified columns are already character type, returns the object unchanged.

This function is particularly useful for handling factor columns that need to be converted to character for downstream analysis functions that expect character input.

Author

Anthony Christidis, anthony-alexander_christidis@hms.harvard.edu